refactor: reorg lib structure for clarity
