+++ lib/map/getIconPathAndColor.tsx
@@ -1,4 +1,4 @@
-import { Category } from "./types";
+import { Category } from "../types";
 
 export const getIconPathAndColor = (
   category?: Category
